Convert RF to DG and RJ to RF
