+++ examples/mcmc/metrohast_norm1d.py
@@ -26,8 +26,8 @@ mu = pb.RV('mu', mu_lims, vtype=float, pscale='log')
 sigma = pb.RV('sigma', sigma_lims, vtype=float, pscale='log')
 x = pb.RV('x', (-np.inf, np.inf), vtype=float)
 sigma.set_mfun((np.log, np.exp))
-paras = pb.RJ(mu, sigma)
-stats = pb.RJ(x)
+paras = pb.RF(mu, sigma)
+stats = pb.RF(x)
 process = pb.SP(stats, paras)
 process.set_prob(scipy.stats.norm.logpdf,
                  order={'x':0, 'mu':'loc', 'sigma':'scale'})
