+++ examples/omc/omc_rs_norm1d.py
@@ -27,9 +27,9 @@ x = pb.RV('x', {-np.inf, np.inf}, vtype=float)
 sigma.set_mfun((np.log, np.exp))
 
 # Set up params and models
-paras = pb.RJ(mu, sigma)
-stats = pb.RJ(x)
-model = pb.RF(stats, paras)
+paras = pb.RF(mu, sigma)
+stats = pb.RF(x)
+model = pb.DG(stats, paras)
 model.set_prob(scipy.stats.norm.logpdf,
                order={'x':0, 'mu':'loc', 'sigma':'scale'},
                pscale='log')
