+++ probayes/__init__.py
@@ -2,8 +2,8 @@ __version__ = '0.0.3.X'
 from probayes.domain import Domain
 from probayes.prob import Prob
 from probayes.rv import RV
-from probayes.rj import RJ
 from probayes.rf import RF
+from probayes.dg import DG
 from probayes.sp import SP
 from probayes.manifold import Manifold
 from probayes.dist import Dist
