+++ probayes/rf_utils.py
@@ -1,32 +1,64 @@
 # Utility module for RF objects
 
+#-------------------------------------------------------------------------------
 import collections
+import numpy as np
+import scipy.stats
+from probayes.pscales import iscomplex, rescale, prod_rule, prod_pscale
 
 #-------------------------------------------------------------------------------
-def desuffix(values, suffix="'"):
-  assert isinstance(values, dict), \
-      "Input must be a dictionary, not {}".format(type(values))
-  suffix_found = any([key[-1] == suffix for key in values.keys()])
-  vals = collections.OrderedDict()
-  if not suffix_found:
-    vals.update({values})
-    return vals
-  for key, val in values.items():
-    vals_key = key if key[-1] != suffix else key[:-1]
-    assert vals_key not in vals, "Repeated key: {}".format(vals_key)
-    vals.update({vals_key: val})
-  return vals
-  
-#-------------------------------------------------------------------------------
-def get_suffixed(values, unsuffix=True, suffix="'"):
-  assert isinstance(values, dict), \
-      "Input must be a dictionary, not {}".format(type(values))
-  vals = collections.OrderedDict()
-  for key, val in values.items():
-    if key[-1] == suffix:
-      vals_key = key[:-1] if unsuffix else key
-      vals.update({vals_key: val})
-  return vals
+def rv_prod_rule(*args, rvs, pscale=None):
+  """ Returns the probability product treating all rvs as independent.
+  Values (=args[0]) are keyed by RV name and rvs are a list of RVs.
+  """
+  values = args[0]
+  pscales = [rv.ret_pscale() for rv in rvs]
+  pscale = pscale or prod_pscale(pscales)
+  use_logs = iscomplex(pscale)
+  probs = [rv.eval_prob(values[rv.ret_name()]) for rv in rvs]
+  prob, pscale = prod_rule(*tuple(probs),
+                           pscales=pscales,
+                           pscale=pscale)
+
+  # This section below is there just to play nicely with conditionals
+  if len(args) > 1:
+    if use_logs:
+      prob = rescale(prob, pscale, 0.j)
+    else:
+      prob = rescale(prob, pscale, 1.)
+    for arg in args[1:]:
+      if use_logs:
+        offs, _ = rv_prod_rule(arg, rvs=rvs, pscale=0.j)
+        prob = prob + offs
+      else:
+        coef, _ = rv_prod_rule(arg, rvs=rvs, pscale=1.)
+        prob = prob * coef
+    if use_logs:
+      prob = prob / float(len(args))
+      prob = rescale(prob, 0.j, pscale)
+    else:
+      prob = prob ** (1. / float(len(args)))
+      prob = rescale(prob, 1., pscale)
+  return prob, pscale
 
 #-------------------------------------------------------------------------------
+def call_scipy_prob(func, pscale, *args, **kwds):
+  index = 1 if iscomplex(pscale) else 0
+  return func[index](*args, **kwds)
 
+#-------------------------------------------------------------------------------
+def sample_cond_cov(*args, cond_cov=None, **kwds):
+    kwds = dict(kwds)
+    cond_pdf = False if 'cond_pdf' not in kwds else kwds.pop('cond_pdf')
+    assert cond_cov, "coveig object mandatory"
+    if len(args) == 1 and isinstance(args[0], dict):
+      vals = args[0]
+      idx = {key: i for i, key in enumerate(vals.keys())}
+      args = [np.array(val) for val in vals.values()]
+    elif not len(args) and len(kwds):
+      vals = dict(kwds)
+      idx = {key: i for i, key in enumerate(vals.keys())}
+      args = list(kwds.values())
+    return cond_cov.interp(*args, cond_pdf=cond_pdf)
+
+#-------------------------------------------------------------------------------
